trailing zeros program simple logic.

+++ src/mathematics/Factorial_Iterative.java
@@ -8,8 +8,8 @@ public class Factorial_Iterative {
 
     }
 
-    public static int factorial (int n) {
-        int factorial = 1;
+    public static long factorial (long n) {
+        long factorial = 1;
         for (int i = 1; i <= n; i++) {
             factorial *= i;
         }
